python-agents/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
python-multipart==0.0.20
beautifulsoup4==4.12.3
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
python-dotenv==1.0.1
